Break up helper.js file back to original locations :/
