+++ lib/rewriter.js
@@ -5,7 +5,6 @@ const ts = require('typescript');
 const tsMorph = require('ts-morph');
 const cleanConsole = require('./mods/clean-console.js');
 const replaceConstructors = require('./mods/replace-constructors.js');
-const {addAllImports} = require('./helpers.js');
 
 const compilerOptions = {
   allowJs: true,
@@ -16,6 +15,21 @@ const compilerOptions = {
   noEmitOnError: false,
   outDir: 'dist', // we're in-memory, but this is still required for some reason
 };
+
+/**
+ * Adds all imports from the list to the top of the source file
+ *
+ * @param {SourceFile} sourceFile - the file to add the imports to
+ * @param {Object} imports - the final list of imports
+ */
+function addAllImports(sourceFile, imports) {
+  for (const [module, originals] of Object.entries(imports)) {
+    sourceFile.addImportDeclaration({
+      defaultImport: originals[0],
+      moduleSpecifier: `std:global/${module}`,
+    });
+  }
+}
 
 /**
  * description
